Done the styling for the grid class and for navigation bar.
I specified that grid container must have grid layout,  defined the columns of the grid along with the gap between columns. Then, I set the position to sticky for navigation bar which should be stick to the top of the view port. Then, set the list items as flex container .