Made changes  for looping through images .
For loop is used to iterate through each element in the image array and set the src attribute of the newly created image to image path in the images directory and alt attribute to the alternative text corresponding to the images from the alternativeTextArray. Also I added click event listener to each thumbnail image.